CLN: remove build some warnings
CLN: from algos_rank_helper.pxi.in, algos_groupby_helper.pxi.in
CLN: hashtable warns
CLN: parser warnings

closes #15190

Author: Jeff Reback <jeff@reback.net>

Closes #15259 from jreback/warn and squashes the following commits:

12cc061 [Jeff Reback] CLN: remove some warnings from algos_rank_helper.pxi.in CLN: and algos_groupby_helper.pxi.in CLN: hashtable warns CLN: parser warnings

+++ pandas/src/parser/io.c
@@ -215,7 +215,7 @@ void *buffer_mmap_bytes(void *source, size_t nbytes, size_t *bytes_read,
 
     retval = src->memmap + src->position;
 
-    if (src->position + nbytes > src->last_pos) {
+    if (src->position + (off_t)nbytes > src->last_pos) {
         // fewer than nbytes remaining
         *bytes_read = src->last_pos - src->position;
     } else {
